Refactor register index types from u32 to usize across various architectures
- Changed register index types from u32 to usize in RISC-V and AArch64 implementations to improve compatibility with pointer arithmetic and indexing.
- Updated function signatures in guest VCPU implementations to accept usize for register indices.
- Adjusted MMIO read/write functions to use usize for register indices.
- Modified the hypervisor types to reflect the changes in register index types.
- Ensured that all related functions and structures are consistent with the new usize type for register indices.

+++ kernel/src/arch/aarch64/hv/guest_vcpu.rs
@@ -91,11 +91,11 @@ impl GuestVcpu {
         unsafe { self.fpu.restore() };
     }
 
-    pub fn get_mmio_data(&self, reg_idx: u8, size: u8) -> u64 {
-        if reg_idx as usize >= self.iregs.len() {
+    pub fn get_mmio_data(&self, reg_idx: usize, size: u8) -> u64 {
+        if reg_idx >= self.iregs.len() {
             return 0;
         }
-        let val = self.iregs[reg_idx as usize];
+        let val = self.iregs[reg_idx];
         match size {
             1 => val & 0xFF,
             2 => val & 0xFFFF,
@@ -104,8 +104,8 @@ impl GuestVcpu {
         }
     }
 
-    pub fn set_mmio_data(&mut self, reg_idx: u8, size: u8, data: u64) {
-        if reg_idx as usize >= self.iregs.len() {
+    pub fn set_mmio_data(&mut self, reg_idx: usize, size: u8, data: u64) {
+        if reg_idx >= self.iregs.len() {
             return;
         }
         let mask = match size {
@@ -114,14 +114,14 @@ impl GuestVcpu {
             4 => 0xFFFFFFFF,
             _ => !0,
         };
-        let old = self.iregs[reg_idx as usize];
+        let old = self.iregs[reg_idx];
         let new = (old & !mask) | (data & mask);
-        self.iregs[reg_idx as usize] = new;
+        self.iregs[reg_idx] = new;
     }
 
-    pub fn get_reg(&self, index: u32) -> Result<u64, &'static str> {
+    pub fn get_reg(&self, index: usize) -> Result<u64, &'static str> {
         match index {
-            reg::X0..=reg::X30 => Ok(self.iregs[index as usize]),
+            reg::X0..=reg::X30 => Ok(self.iregs[index]),
             reg::SP => Ok(self.sysregs.sp_el1),
             reg::PC => Ok(self.pc),
             reg::PSTATE => Ok(self.spsr),
@@ -147,10 +147,10 @@ impl GuestVcpu {
         }
     }
 
-    pub fn set_reg(&mut self, index: u32, value: u64) -> Result<(), &'static str> {
+    pub fn set_reg(&mut self, index: usize, value: u64) -> Result<(), &'static str> {
         match index {
             reg::X0..=reg::X30 => {
-                self.iregs[index as usize] = value;
+                self.iregs[index] = value;
                 Ok(())
             }
             reg::SP => {
